moved messages into hook
