+++ packages/app/features/chat/chat-screen.tsx
@@ -4,62 +4,19 @@ import { View } from 'app/design/view'
 import { GiftedChat } from 'react-native-gifted-chat'
 
 import { createParam } from 'solito'
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback } from 'react'
 import { useCommunity } from 'app/hooks/useCommunity'
+import { IMessage, useRoomMessages } from 'app/hooks/useRoomMessages'
 
 const { useParam } = createParam<{ id: string }>()
-export interface IMessage {
-  _id: string | number
-  text: string
-  createdAt: Date | number
-  user: {_id: string | number, name?: string, avatar?: string }
-  image?: string
-  video?: string
-  audio?: string
-  system?: boolean
-  sent?: boolean
-  received?: boolean
-  pending?: boolean
-  quickReplies?: QuickReplies
-}
-interface Reply {
-  title: string
-  value: string
-  messageId?: any
-}
-
-interface QuickReplies {
-  type: 'radio' | 'checkbox'
-  values: Reply[]
-  keepIt?: boolean
-}
 
 export function ChatScreen() {
   const [id] = useParam('id')
   const { getRoomById } = useCommunity()
   const room = getRoomById(id)
+  const [messages, setMessages] = useRoomMessages(id)
 
-  const [messages, setMessages] = useState<IMessage[]>([]);
-
-  useEffect(() => {
-    const messages: IMessage[] = []
-    for (let i = 0; i < 20; i++) {
-      const message:IMessage = {
-        _id: i,
-        text: 'This is message ' + ( i + 1),
-        createdAt: new Date(Date.now()-(i*3600000)),
-        user: {
-          _id: 2,
-          name: 'React Native',
-          avatar: 'https://placeimg.com/140/140/any',
-        },
-      }
-      messages.push(message);
-    }
-    setMessages(messages)
-  }, [])
-
-  const onSend = useCallback((messages:IMessage[] = []) => {
+  const onSend = useCallback((messages: IMessage[] = []) => {
     setMessages(previousMessages => GiftedChat.append(previousMessages, messages))
   }, [])
